feat(modules/mobile-tests): review remarks
