+++ modules/mobile-tests/specs/test.e2e.js
@@ -3,7 +3,7 @@
 const wdio = require('webdriverio');
 
 const {
-    ONEPLUSA5000
+    ANDROID
 } = require('../helpers/capabilities');
 const { getSelector } = require('../helpers/getSelector');
 
@@ -11,7 +11,7 @@ const { getSelector } = require('../helpers/getSelector');
 const options = {
     path: '/wd/hub',
     port: 4723,
-    capabilities: ONEPLUSA5000
+    capabilities: ANDROID
 };
 
 const delay = ms => new Promise(res => setTimeout(res, ms));
