+++ package.json
@@ -61,7 +61,7 @@
     "@vladmandic/human": "2.6.5",
     "@vladmandic/human-models": "2.5.9",
     "@wdio/sync": "7.20.9",
-    "@xmldom/xmldom": "0.7.5",
+    "@xmldom/xmldom": "0.7.6",
     "amplitude-js": "8.2.1",
     "appium": "1.22.3",
     "appium-doctor": "1.16.0",
@@ -193,7 +193,6 @@
     "ts-loader": "9.4.1",
     "typescript": "4.7.4",
     "unorm": "1.6.0",
-    "wdio-chromedriver-service": "8.0.0",
     "webpack": "5.57.1",
     "webpack-bundle-analyzer": "4.4.2",
     "webpack-cli": "4.9.0",
@@ -215,7 +214,7 @@
     "postinstall": "patch-package --error-on-fail && jetify",
     "validate": "npm ls",
     "start": "make dev",
-    "wdio": "wdio run wdio.conf.js --watch"
+    "wdio": "wdio run modules/mobile-tests/wdio.conf.js --watch"
   },
   "resolutions": {
     "@types/react": "17.0.14",
